+++ src/index.ts
@@ -3,3 +3,5 @@ export { default as FhiButton } from './components/FhiButton';
 export { default as FhiDetails } from './components/FhiDetails';
 export { default as FhiTable } from './components/FhiTable';
 export { default as FhiHeader } from './components/FhiHeader';
+export { default as FhiTabBar } from './components/FhiTabBar';
+export { default as FhiAccordion } from './components/FhiAccordion';